7-8 done

+++ app/Http/Controllers/ArticleController.php
@@ -2,9 +2,13 @@
 
 namespace App\Http\Controllers;
 
-use Illuminate\Http\Request;
 use App\Models\Article;
 use App\Models\Comment;
+use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
+use Illuminate\Support\Facades\Mail;
+use Illuminate\Support\Facades\Log;
+use App\Jobs\ArticleMailJob;
 
 class ArticleController extends Controller
 {
@@ -16,12 +20,17 @@ class ArticleController extends Controller
 
     public function store()
     {
-        $data = request() -> validate([
+        request() -> validate([
             'title' => 'required|string|max:255',
             'desc' => 'required|string',
         ]);
 
-        Article::create($data);
+        $article = new Article;
+        $article->title = request()->title;
+        $article->desc = request()->desc;
+        $article->author_id = 1;
+        $article->save();
+        ArticleMailJob::dispatch($article);
 
         return redirect()->route('articles');
     }
@@ -33,6 +42,7 @@ class ArticleController extends Controller
 
     public function edit(Article $article)
     {
+        Gate::authorize('update', [self::class, $article]);
         return view("articles.edit", compact("article"));
     }
 
@@ -44,12 +54,15 @@ class ArticleController extends Controller
         ]);
         $article->title = request() -> get('title', '');
         $article->desc = request() -> get('desc', '');
+        $article->author_id = 1;
         $article->save();
         return redirect()->route('articles');
     }
 
     public function destroy(Article $article)
     {
+        Gate::authorize('delete', [self::class, $article]);
+        
         $article->delete();
 
         return redirect()->route('articles');
